+++ Academia/WebApi.DTO/Mappers/PersonaMapper.cs
@@ -1,7 +1,7 @@
 ﻿using Domain.Entities;
 using WebApi.DTO.Dto;
 using WebApi.DTO.Request;
-using WebApi.DTO.Respone;
+using WebApi.DTO.Response;
 
 namespace WebApi.DTO.Mappers
 {
