+++ Academia/WebApi.DTO/Response/PersonaResponse.cs
@@ -5,7 +5,7 @@ using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WebApi.DTO.Respone
+namespace WebApi.DTO.Response
 {
     public class PersonaResponse
     {
