+++ Academia/WebApi/Endpoints/Persona/Create.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Services.Contracts;
 using WebApi.DTO.Mappers;
 using WebApi.DTO.Request;
-using WebApi.DTO.Respone;
+using WebApi.DTO.Response;
 
 namespace WebApi.Endpoints.Persona
 {
